+++ index.html
@@ -7,7 +7,9 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Alan Balen Schio - @schirrel</title>
   <link href="https://fonts.googleapis.com/css?family=ZCOOL+QingKe+HuangYou" rel="stylesheet" />
+  <link rel="stylesheet" href="reset.css">
   <link rel="stylesheet" href="maquiagem.css">
+
   <script data-ad-client="ca-pub-5684526463125026" async
     src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script>
 </head>
@@ -16,15 +18,22 @@
 <body>
 
   <header>
-    <span class="logo">ABS</span>
+    <div class="logo">
+      <a href="#start">
+        <span>Alan</span>
+        <span>Balen</span>
+        <span>Schio</span>
+      </a>
+    </div>
     <nav>
       <ol>
-        <li>
-          <a href="#about"  data-i18n="whothis">Quem é esse doido?</a></li>
+        <li><a href="#who" data-i18n="whothis">Sobre</a></li>
+        <li><a href="#what" data-i18n="tricks">Meus Truques</a></li>
+        <li><a href="#when" data-i18n="xp">Experiência</a></li>
+        <li><a href="#wow" data-i18n="contact">Contato</a></li>
       </ol>
     </nav>
     <div class="dropdown">
-
       <button type="button">
         <svg xmlns="http://www.w3.org/2000/svg" height="36" viewBox="0 0 27 27" width="36">
           <path
@@ -39,7 +48,8 @@
           <path
             d="M21.512 11.666h-7.656c-1.208 0-2.19.982-2.19 2.19v4.16l-1.984 1.416a.595.595 0 000 .97l2.008 1.433a2.194 2.194 0 002.166 1.867h7.656c1.208 0 2.19-.982 2.19-2.19v-7.656c0-1.208-.982-2.19-2.19-2.19zm0 0"
             fill="var(--secondary)" stroke="#000" stroke-width=".59568678" />
-          <path d="M21.512 11.666h-3.828v12.036h3.828c1.208 0 2.19-.982 2.19-2.19v-7.656c0-1.208-.982-2.19-2.19-2.19zm0 0"
+          <path
+            d="M21.512 11.666h-3.828v12.036h3.828c1.208 0 2.19-.982 2.19-2.19v-7.656c0-1.208-.982-2.19-2.19-2.19zm0 0"
             fill-opacity=".4" />
           <path
             d="M15.907 19.69c0-.012.007-.038.013-.063l1.231-4.013c.058-.192.294-.281.53-.281.242 0 .478.09.536.28l1.231 4.014c.006.025.013.044.013.064 0 .197-.3.344-.523.344-.14 0-.25-.045-.281-.166l-.242-.848h-1.462l-.242.848c-.032.121-.14.166-.28.166-.224 0-.524-.14-.524-.344zm2.329-1.32l-.555-1.959-.555 1.959zm0 0"
@@ -49,17 +59,91 @@
               d="M11.22 8.863a.336.336 0 000-.672H9.726v-.816a.336.336 0 00-.293-.332v1.82h.648a2.534 2.534 0 01-.648 1.38v.953a3.194 3.194 0 001.787.546.336.336 0 000-.672c-.476 0-.921-.132-1.302-.36.46-.497.765-1.138.84-1.847zm0 0M19.448 19.627l-1.231-4.013c-.057-.19-.292-.28-.533-.281v1.09l.552 1.947h-.552v.65h.73l.243.85c.032.12.14.165.28.165.224 0 .524-.147.524-.344 0-.02-.007-.039-.013-.064zm0 0" />
           </g>
         </svg>
-        </button>
-        <ul id="languages">
-        </ul>
+      </button>
+      <ul id="languages">
+      </ul>
     </div>
   </header>
-  <main></main>
+  <section id="start" class="hero">
+    <article>
+
+      <em>Aquele que quer mudar o mundo usando CSS!</em>
+    </article>
+  </section>
+
+  <section id="who">
+    <span class="title" data-i18n="whothis">Quem é esse doido?</span>
+    <article>
+      <h2>Oie, eu sou o Alan.</h2>
+      <p>Trabalho como Desenvolvedor FullStack focado em Front End, sou apaixonado por design, UI/UX e open source.</p>
+
+
+    </article>
+  </section>
+
+  <section id="what">
+    <span class="title" data-i18n="tricks">Meus Truques </br>
+      <em><strong> Em algumas coisas eu sou bom, em outras sou melhor ainda.
+        </strong>
+      </em>
+
+    </span>
+    <article class="full-width">
+      <div class="flex-row">
+        <div class="flex-col">
+          <strong>Front End</strong>
+          <dl>
+            <dt>CSS </dt>
+            <dd data-tooltip="Sem pré processor nem css-in-js, sim aquele que mete medo em todo mundo">PURO! </dd>
+            <dt>SVG </dt>
+            <dd data-tooltip="A era das gifs acabou! CSS+SVG são o jeito mais lindo e performático de deixar seu desenvolvimento com animações bonitas e eficientes.">Icones e Animações. </dd>
+             <dt>JavaScript</dt>
+            <dd>AngularJS, VueJS, ES6, WebComponents, VanillaJS. </dd>
+            <dt>HTML Semantico</dt>
+            <dd>Além de ser a maneira mais eficiente para o navegador é també/m o jeito de deixar as coisas mais acessiveis.</dd>
+            <dt>UI e UX</dt>
+            <dd><em>"Deixa a web bonita bem, e deixa a web com usabilidade"</em> - <strong>Vovó Juju</strong></dd>
+         </dl>
+        </div>
+        
+        <div class="flex-col">
+          <strong>Back End / Desktop</strong>
+          <ul>
+            <dt>Java, JavaEE, Spring, Vraptor</dt>
+            <dt>NodeJS, AdonisJS</dt>
+            <dt>C++</dt>
+            <dt>C#</dt>
+          </ul>
+        </div>
+        
+        <div class="flex-col">
+          <strong>Outros</strong>
+          <ul>
+            <dt>Linux, A<dt>
+            <dt>SQL, MSSQL, PostgreSQL</dt>
+            <dt>Bitbucket Pipeline</dt>
+            <dt>Inkscape+GIMP</dt>
+            <dt>Apresentações</dt>
+            <dt>+- React, +-Angular, +- TypeScript</dt>
+          </ul>
+        </div>
+      </div>
+    </article>
+  </section>
+
+  <section id="when">
+    <h3 data-i18n="xp">Experiência</h3>
+  </section>
+
+  <section id="wow">
+    <h3 data-i18n="contact">Contato</h3>
+  </section>
+  <section></section>
 
   <footer>
     Made with Love and FlatIcon
   </footer>
-  <script src="magica.js"></script>
+  <script type="module" src="magica.js"></script>
 </body>
 
 </html>
