+++ tooltip.css
@@ -1,0 +1,46 @@
+[data-tooltip]:hover:after, [data-tooltip]:hover:before {
+    visibility: visible;
+    pointer-events: auto;
+}
+[data-tooltip]:after, [data-tooltip]:before {
+    pointer-events: none;
+    visibility: hidden;
+}
+
+[data-tooltip]:after {
+ 
+    pointer-events: none;
+    content: attr(data-tooltip);
+    position: absolute;
+    text-transform: none;
+    text-align: left;
+    white-space: initial;
+    font-size: 1rem;
+    word-break: break-word;
+    border: 1px solid var(--primary);
+    line-height: 1.4285em;
+    max-width: 400px;
+    background: var(--primary);
+    padding: .833em 1em;
+    font-weight: 400;
+    font-style: normal;
+    color: rgba(0,0,0,.87);
+    border-radius: .28571429rem;
+    -webkit-box-shadow: 0 2px 4px 0 rgba(34,36,38,.12), 0 2px 10px 0 rgba(34,36,38,.15);
+    box-shadow: 0 2px 4px 0 rgba(34,36,38,.12), 0 2px 10px 0 rgba(34,36,38,.15);
+    z-index: 1;
+    margin-top: -25px;
+    -webkit-transform: translateX(-500%) scale(0);
+    transform: translateX(-500%) scale(0);
+    transition: all 300ms ease;
+}
+
+[data-tooltip]:hover:after, [data-tooltip][data-position="bottom center"]:hover:after {
+
+    -webkit-transform: translateX(0%) scale(1)!important;
+    transform: translateX(0%) scale(1)!important;
+}
+
+[data-tooltip] {
+    cursor: help;
+}
